Refactor [#9] 홈 오픈 예정 뷰 분리
